Add Packer templates

+++ tmp/base.sh
@@ -5,12 +5,13 @@ set -e
 echo "Configuring the portage environment..."
 mv /tmp/make.conf /etc/portage/make.conf
 grep --quiet "MAKEOPTS=" /etc/portage/make.conf || sed --in-place "/CPU_FLAGS_X86=/a MAKEOPTS=\"-j$(expr $(nproc) + 1)\"" /etc/portage/make.conf
-mkdir -p /usr/portage/{distfiles,metadata,packages}
-echo "masters = gentoo" > /usr/portage/metadata/layout.conf
-chown -R portage:portage /usr/portage
 
 echo "Installing a portage snapshot..."
 emerge-webrsync --quiet
+
+echo "Purging news items..."
+eselect news read --quiet
+eselect news purge
 
 echo "Updating the portage tree..."
 emerge --sync --quiet
@@ -34,15 +35,12 @@ echo "Reloading the environment..."
 env-update && source /etc/profile
 
 echo "Updating the world..."
+find / -type f \( -perm -4000 -or -perm -2000 \) -links +1 ! -path "/proc/*" -exec chmod u-s,g-s {} \;
 emerge --emptytree --update --deep --newuse --quiet @world
 
 echo "Updating configuration files..."
 etc-update --preen --quiet
 find /etc/ -name "._cfg[0-9][0-9][0-9][0-9]*" -exec rm {} \;
-
-echo "Purging news items..."
-eselect news read --quiet
-eselect news purge
 
 echo "Purging unused locales..."
 emerge --quiet localepurge
@@ -53,16 +51,21 @@ echo "Cleaning the source files directory..."
 emerge --quiet gentoolkit
 eclean-dist --deep --quiet
 
+echo "Packaging development tools..."
+quickpkg --include-config y autoconf automake bison yacc binutils libtool gcc localepurge libltdl autoconf-wrapper binutils-config help2man automake-wrapper gcc-config mpc mpfr gmp
+
 echo "Cleaning the system..."
 emerge --unmerge --quiet autoconf automake bison yacc binutils libtool gcc localepurge
 emerge --depclean --quiet
-echo "nameserver 8.8.8.8" > /etc/resolv.conf
-echo "nameserver 8.8.4.4" >> /etc/resolv.conf
+if [ -f /etc/resolv.conf ]
+then
+	echo "nameserver 8.8.8.8" > /etc/resolv.conf
+	echo "nameserver 8.8.4.4" >> /etc/resolv.conf
+fi
 mkdir --parents /usr/local/bin
 mv /tmp/portagepurge /usr/local/bin/portagepurge
 chmod +x /usr/local/bin/portagepurge
 portagepurge
-rm --force /etc/resolv.conf
 
 echo "Setting the default options for emerge..."
 grep --quiet "EMERGE_DEFAULT_OPTS=" /etc/portage/make.conf || sed -i '/FEATURES=/a EMERGE_DEFAULT_OPTS="--usepkg --buildpkg"' /etc/portage/make.conf
